Create CONTRIBUTING.md
If you want to add new Notebooks to this project, please follow these steps:

Create a new Notebook: The name of the notebook should include the problem number with leading zeros. This ensures that a string search shows problem 0001 before 1000.

Solve the Problem: Write markdown and code in the Notebook to solve the challenge. Make sure to add comments that emphasize the relevant topics from the README.

(Future feature -- do not try this yet) Update the Changelog: Once the Notebook has run and provided a solution, update the Unreleased section of the CHANGELOG. Mention that the solution number has been solved. If applicable, add up to three one-sentence sub-bullets highlighting what lessons learned from the README are prominent in this solution.

Shut Down the Kernel: After you're done solving the challenge, shut down the Kernel for the Notebook. This helps to free up resources for other tasks.